+++ interpreter/main.c
@@ -280,12 +280,12 @@ tremain_print_hardinfo (void)
 	printf ("Max. symbol length: %d characters\n", TRE_MAX_SYMLEN);
 	printf ("Max. literal string length: %d characters\n", TRE_MAX_STRINGLEN);
 	printf ("Max. packages: %d\n", MAX_PACKAGES);
-	printf ("Cells start: %8lX\n", (ulong) &tre_lists);
-	printf ("Cells end:   %8lX\n", (ulong) &tre_lists[NUM_LISTNODES]);
-	printf ("Atoms start: %8lX\n", (ulong) &tre_atoms);
-	printf ("Atoms end:   %8lX\n", (ulong) &tre_atoms[NUM_ATOMS]);
-	printf ("Nums start:  %8lX\n", (ulong) &tre_numbers);
-	printf ("Nums end:    %8lX\n", (ulong) &tre_numbers[NUM_NUMBERS]);
+	printf ("Cells start: %8lX\n", (size_t) &tre_lists);
+	printf ("Cells end:   %8lX\n", (size_t) &tre_lists[NUM_LISTNODES]);
+	printf ("Atoms start: %8lX\n", (size_t) &tre_atoms);
+	printf ("Atoms end:   %8lX\n", (size_t) &tre_atoms[NUM_ATOMS]);
+	printf ("Nums start:  %8lX\n", (size_t) &tre_numbers);
+	printf ("Nums end:    %8lX\n", (size_t) &tre_numbers[NUM_NUMBERS]);
 
 	printf ("Max. files: %d\n", TRE_FILEIO_MAX_FILES);
 	printf ("Max. nested files (interpreter): %d\n", TRE_MAX_NESTED_FILES);
